refactor(kvm): use strongly-typed registers in KVM bindings

+++ alioth/src/hv/kvm/vcpu/vcpu_x86_64/vcpu_x86_64_test.rs
@@ -280,10 +280,10 @@ fn test_kvm_run() {
     let idtr = DtRegVal { base: 0, limit: 0 };
     vcpu.set_sregs(
         &[
-            (SReg::Efer, (Efer::LMA | Efer::LME).bits() as u64),
-            (SReg::Cr0, (Cr0::NE | Cr0::PE | Cr0::PG).bits() as u64),
+            (SReg::Efer, (Efer::LMA | Efer::LME).bits()),
+            (SReg::Cr0, (Cr0::NE | Cr0::PE | Cr0::PG).bits()),
             (SReg::Cr3, 0x2000),
-            (SReg::Cr4, Cr4::PAE.bits() as u64),
+            (SReg::Cr4, Cr4::PAE.bits()),
         ],
         &[
             (SegReg::Cs, cs),
